Use cast to remove a type ignore comment
